feat: more

+++ runtcc.sh
@@ -21,23 +21,32 @@ check_existing_ami() {
 
 run_packer() {
     echo "Entering Packer"
+    if [ -f THAT_IMAGE ]; then
+       cluster_ami=$(cat THAT_IMAGE)
+       return
+    fi
     echo "Running Packer..."
     cd packer
     packer build main.pkr.hcl -machine-readable | tee packer.log
     cluster_ami=$(tail -n1 packer.log | grep -o 'ami-[0-9a-f]*')
     rm packer.log
     cd ..
+    echo $cluster_ami>THAT_IMAGE
     echo "Exiting Packer"
 }
 
 run_terraform() {
     echo "Entering Terraform"
     cd terraform
-    terraform destroy -auto-approve
+#    terraform destroy -auto-approve
     terraform apply -auto-approve -var "cluster_ami=${cluster_ami}"
     PUBLIC_HOSTS=$(terraform output -json | jq '.public_ip_addresses.value' | grep \" | sed 's/"//g' | sed 's/,//g' | sed 's/ //g' | paste -s -d ',')
     PRIVATE_HOSTS=$(terraform output -json | jq '.private_dns.value' | grep \" | sed 's/"//g' | sed 's/,//g' | sed 's/ //g' | paste -s -d ',')
     HOSTS_SIZE=$(terraform output -json | jq '.private_dns.value | length' )
+    if [ $HOSTS_SIZE -eq 1 ]; then
+	PRIVATE_HOSTS=$PRIVATE_HOSTS,
+	PUBLIC_HOSTS=$PUBLIC_HOSTS,
+    fi
     echo PRIVATE_HOSTS=${PRIVATE_HOSTS}
     LUSTRE_DNS_NAME=$(terraform output -json | jq '.lustre_dns_name.value')
     LUSTRE_MOUNT_NAME=$(terraform output -json | jq '.lustre_mount_name.value')
@@ -61,6 +70,7 @@ main() {
     echo firstMachine = $firstMachine
     run_ansible
     bash -x ./runMpi.sh $firstMachine $PRIVATE_HOSTS $HOSTS_SIZE
+#    bash -x ./runMpi.sh $firstMachine $PUBLIC_HOSTS $HOSTS_SIZE
 }
 
 main
